GQAnalyzer: A package for Hydrogeochemical plotting

This package provides functions to create the following plots:

Piper plot

Stiff diagrams

Durov plot

Ternary

Radial

Multirectangular

Maucha

Anomaly Utilities

This section includes functions used in the identification of anomalies in groundwater quality data.

This section includes

Durov Diagram

The Durov diagram is a composite diagram that includes a central rectangular diagram of

This section includes the definition of three functions used to create the Durov diagram: durov_transform, durov_transform1, ggplot_durov

Maucha Diagram

The functions defined to create a Multirectangular diagram are: maucha_transform, ggplot_maucha

Mixing Utilities

This section is devoted to functions used in the study of mixing processes from hydrogeochenmical information.

This section includes the definition of three functions used to create the Durov diagram: m3_model

Multirectangular Diagram

The functions defined to create a Multirectangular diagram are: multirectangular_transform, ggplot_multirectangular

Piper Diagram

The functions defined to create a Piper diagram are: anion_cation_tranform, transform_piper_data, ggplot_piper

Radial Diagram

The functions defined to create a Radial diagram are: radial_tranform, ggplot_radial

Stiff Diagram

The functions defined to create a Stiff diagram are: stiff_tranform, ggplot_stiff

Ternary Plot

The functions defined to create a Ternary plot are: ternary_transform, ggplot_ternary