geochemical_dataset

Function to create a geochemical_dataset.

geochemical_dataset(name = "GeochemicalDataset", data)

Arguments

name
Name of the geochemical dataset

data
Data.frame with the concentration of anions and cations (see details)

Value

An object of the geochemical_dataset class with the following member variables:

name: Name of the geochemical dataset

dataset: data.frame with the corresponding measured concentrations

meqL: data.frame with the meqL of the measured concentrations

anions: data.frame with the meqL of the major anions

cations: data.frame with the meqL of the major cations

meqL.min: Numeric vector with the minimum values of the meqL

meqL.max: Numeric vector with the maximum values of the meqL

meqL.schoeller: data.frame with the concentrations of the major ions organized to create tehe Schoeller plot.

Details

A geochemical dataset is a data.frame with the values of the concentrations of the anions and cations obtained during hydrogeochemical sampling. Internally this object calculates the meqL of the measured concentrations of the major ions. The concentrations of the ions must be specified in mg/L.

See also

Other base functions: convert_meql, plot.geochemical_dataset