plot.geochemical_dataset

Generic function to create different hydrogeochemical plots

# S3 method for geochemical_dataset
plot(x, y = NULL, ..., type = c("piper",
  "ternary", "durov", "schoeller", "stiff", "multirectangular", "radial"),
  measure = c("conc", "meql"), vars = NULL, color = NULL,
  Size = NULL)

Arguments

x, y
A geochemical_dataset object

...
Additional parameters to the plot function

type
A character string with the type of hydrogeochemical plot to be created. Currently the following plots are supported:

piper

ternary

durov

schoeller

stiff

multirectangular

radial

maucha

measure
A character variable specifying the type of measure to be used in the plot. Currently the types supported include:

'conc': concentrations

'meql': miliequivalent

vars
Character vector. Variables to use in the plot. In the case of the Durov plot, these variables are already defined. Used only for compatibility with the plot function.

color
Character variable that specifies the variable to color the data inside the plot.

Size
Character variable that specifies the variable to define the size of the data inside the plot.

Value

This function returns a ggplot2 object with the corresponding hydrogeochemical plot.

See also