plot_radial

Function to create a radial plot from a single sample

plot_radial(x, measure = c("conc", "meql"), vars = NULL, color = NULL,
  Size = NULL)

Arguments

x
A geochemical_dataset object

measure
A character variable specifying the type of measure to be used in the plot. Currently the types supported include:

'conc': concentrations

'meql': miliequivalent

vars
Character vector. Variables to use in the plot. In the case of the Durov plot, these variables are already defined. Used only for compatibility with the plot function.

color
Character variable that specifies the variable to color the data inside the plot.

Size
Character variable that specifies the variable to define the size of the data inside the plot.

Value

This function returns a ggplot2 object with the Durov plot.

See also

Other radial functions: ggplot_radial, radial_transform